Add getUserFeed function for authenticated user's feed retrieval
